Save trade strand into db; Remove trade strand from quota repository form; Fix FK issue at spider side; Generate Trade Strand faker;

+++ IntelligentInvestor.Domain/Trades/TradeStrand.cs
@@ -16,41 +16,41 @@ public class TradeStrand : QuotaBase
 
     public long BuyStrand1 { get; set; }
 
-    public double BuyPrice1 { get; set; }
+    public decimal BuyPrice1 { get; set; }
 
     public long BuyStrand2 { get; set; }
 
-    public double BuyPrice2 { get; set; }
+    public decimal BuyPrice2 { get; set; }
 
     public long BuyStrand3 { get; set; }
 
-    public double BuyPrice3 { get; set; }
+    public decimal BuyPrice3 { get; set; }
 
     public long BuyStrand4 { get; set; }
 
-    public double BuyPrice4 { get; set; }
+    public decimal BuyPrice4 { get; set; }
 
     public long BuyStrand5 { get; set; }
 
-    public double BuyPrice5 { get; set; }
+    public decimal BuyPrice5 { get; set; }
 
     public long SellStrand1 { get; set; }
 
-    public double SellPrice1 { get; set; }
+    public decimal SellPrice1 { get; set; }
 
     public long SellStrand2 { get; set; }
 
-    public double SellPrice2 { get; set; }
+    public decimal SellPrice2 { get; set; }
 
     public long SellStrand3 { get; set; }
 
-    public double SellPrice3 { get; set; }
+    public decimal SellPrice3 { get; set; }
 
     public long SellStrand4 { get; set; }
 
-    public double SellPrice4 { get; set; }
+    public decimal SellPrice4 { get; set; }
 
     public long SellStrand5 { get; set; }
 
-    public double SellPrice5 { get; set; }
+    public decimal SellPrice5 { get; set; }
 }
